:truck: rename components to lowercase
