+++ src/components/register/background/index.tsx
