+++ src/components/register/background/styles.module.scss
