+++ src/components/register/form/screens/education-p1.tsx
