+++ src/components/register/form/screens/education-p2.tsx
