+++ src/components/register/form/screens/welcome.tsx
