+++ src/components/register/navigation/styles.module.scss
